TMP: Ajuste en SuperModelo Usuario, prueba de listados con SubModelo Profesor

+++ application/controllers/profesor/Profesor.php
@@ -21,18 +21,20 @@ class Profesor extends CI_Controller {
      */
     public function __construct() {
         parent::__construct();
-        if ($this->session->userdata("id_rol_usuario") != ID_ROL_ADMINISTRADOR || $this->user_model->isLoggedIn() !== TRUE) {
-            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
-            redirect('user/login');
-        }
         $this->load->model('profesor_model');
         $this->load->helper('html_builder_helper');
     }
 
     public function index() {
-        $lista_profesores = $this->profesor_model->listarAdmin();
+
+        if ($this->user_model->isLoggedIn() !== TRUE) {
+            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
+            redirect('user/login');
+        }
+
+        $lista_profesores = $this->profesor_model->getAll();
         //die(print_r($lista_profesores,true));
-        $html = usuario_list_table($lista_profesores, 'profesor');
+        $html = usuario_list_table($lista_profesores,'profesor');
 
         $data ["titulo"] = "Lista de profesores";
         $data ["html"] = $html;
@@ -40,16 +42,28 @@ class Profesor extends CI_Controller {
     }
 
     public function view($id) {
+        if ($this->user_model->isLoggedIn() !== TRUE) {
+            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
+            redirect('user/login');
+        }
+        
         $this->load->model("facultades_model");
         $this->load->model("sedes_model");
-
-        $data["profesor"] = get_object_vars($this->profesor_model->obtener($id)[0]);
+        
+        $data["profesor"] = get_object_vars($this->profesor_model->get($id)[0]);
         $data ["titulo"] = "Detalles de un profesor - SEPP";
-
+        
         $this->load->view("admin/profesor/view", $data);
+        
     }
 
     public function add() {
+
+        if ($this->user_model->isLoggedIn() !== TRUE) {
+            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
+            redirect('user/login');
+        }
+
         $this->load->model("facultades_model");
         $this->load->model("sedes_model");
 
@@ -83,10 +97,16 @@ class Profesor extends CI_Controller {
     }
 
     public function edit($id = "") {
+
+        if ($this->user_model->isLoggedIn() !== TRUE) {
+            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
+            redirect('user/login');
+        }
+
         $this->load->model("facultades_model");
         $this->load->model("sedes_model");
         $datosProfesor = $this->profesor_model->get($id);
-        if ($datosProfesor == NULL) {
+        if($datosProfesor == NULL){
             redirect('admin/profesor', 'refresh');
         }
         $data["sedes"] = $this->sedes_model->SelectAllSedes();
@@ -118,6 +138,11 @@ class Profesor extends CI_Controller {
     }
 
     public function remove($id) {
+
+        if ($this->user_model->isLoggedIn() !== TRUE) {
+            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
+            redirect('user/login');
+        }
         if ($this->input->is_ajax_request()) {
             $this->profesor_model->delete(['id' => $id]);
             $this->session->set_flashdata('message', "Usuario deshabilitado exitosamente.");
@@ -127,8 +152,13 @@ class Profesor extends CI_Controller {
             redirect('admin/profesor');
         }
     }
-
+    
     public function enable($id) {
+
+        if ($this->user_model->isLoggedIn() !== TRUE) {
+            $this->session->set_flashdata('error', "Debe autenticarse para ingresar a &eacute;sta opci&oacute;n.");
+            redirect('user/login');
+        }
         if ($this->input->is_ajax_request()) {
             $this->profesor_model->enable(['id' => $id]);
             $this->session->set_flashdata('error', "Usuario habilitado exitosamente.");
